resend project user invitation if deleted or expired
If an invitation was sent but not accepted by the project user for some times, it expires, and on the next terraform apply, there is an error saying that the invitation that is already in the state is not found. To avoid this, we are sending an invitation once more if it was deleted or expired (we cannot tell which is the case) but only when the invitation was not accepted before.