simplify handling of preexisting service integrations
currently it is allowed to import arbitary service integrations which is
non canonical behaviour. preexisting resources should be imported as opt-in.
This PR restricts the allowed preexisting integrations to be of `read_replica`
type since that is the only type that can be created alongside a service
by terraform itself.