Add deprecation notice to documentation of the `aiven_service_user` resource
Today during a code review I noticed someone using this resource. When trying to find a reference of the deprecation it is only mentioned in the release notes and in the runtime deprecation warning, I would expect this to be present in the documentation of the resource too.

Adding a warning box for this, copying most of the phrasing from the runtime deprecation warning.